+++ src/PizzaOrderPreview.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const PizzaOrderPreview = ({pizza}) => {
+  return <div>Pizza</div>;
+};
